Added functionality to fetch and save the user's TOTP code to storage
