+++ PennMobile/Login/LoginWebviewController.swift
@@ -47,9 +47,6 @@ class LoginWebviewController: PennLoginController, IndicatorEnabled {
                     self.loginCompletion?(false)
                 }
                 UserDefaults.standard.storeCookies()
-                
-                // Create TOTP App
-//                TOTPFetcher.instance.fetchAndSaveTOTPSecret()
             }
         }
     }
@@ -131,7 +128,6 @@ class LoginWebviewController: PennLoginController, IndicatorEnabled {
                 self.getRemainingCourses()
                 self.getDiningBalance()
                 self.getDiningTransactions(after: 0.5)
-                TOTPFetcher.instance.fetchAndSaveTOTPSecret()
                 
                 if accountID == nil {
                     FirebaseAnalyticsManager.shared.trackEvent(action: "Attempt Login", result: "Failed Login", content: "Failed Login")
